+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/adapter/SplittableRandomAdapterDemoTest.java
@@ -3,12 +3,12 @@ package io.github.pr0methean.betterrandom.prng.adapter;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import org.testng.annotations.Test;
 
-public class ReseedingSplittableRandomAdapterDemoTest {
+public class SplittableRandomAdapterDemoTest {
 
   private static final String[] NO_ARGS = {};
 
   @Test(timeOut = 120_000) public void ensureNoDemoCrash()
       throws SeedException, InterruptedException {
-    ReseedingSplittableRandomAdapterDemo.main(NO_ARGS);
+    SplittableRandomAdapterDemo.main(NO_ARGS);
   }
 }